# [PERF] Route result should be of a fixed size