--- a/lib/route-recognizer.js
+++ b/lib/route-recognizer.js
@@ -294,6 +294,8 @@ function findHandler(state, path, queryParams) {
   var captures = path.match(regex), currentCapture = 1;
   var result = new RecognizeResults(queryParams);
 
+  result.length = handlers.length;
+
   for (var i=0; i<handlers.length; i++) {
     var handler = handlers[i], names = handler.names, params = {};
 
@@ -301,7 +303,7 @@ function findHandler(state, path, queryParams) {
       params[names[j]] = captures[currentCapture++];
     }
 
-    result.push({ handler: handler.handler, params: params, isDynamic: !!names.length });
+    result[i] = { handler: handler.handler, params: params, isDynamic: !!names.length };
   }
 
   return result;