Speeds up recognize by about 60,000 op/s. This is w/o #77.

```
  Add ................... 519.78 op/s
  End-to-end ............ 261.00 op/s
  Generate ........ 4,822,764.15 op/s
  Handlers For ... 37,881,724.23 op/s
  Recognize ......... 426,530.12 op/s
```